address CR comments
