+++ packages/flutter/lib/scheduler.dart
@@ -15,6 +15,5 @@ library scheduler;
 
 export 'src/scheduler/binding.dart';
 export 'src/scheduler/debug.dart';
-export 'src/scheduler/performance_mode.dart';
 export 'src/scheduler/priority.dart';
 export 'src/scheduler/ticker.dart';
