+++ packages/flutter/test/scheduler/performance_mode_test.dart
@@ -5,58 +5,58 @@
 import 'dart:ui';
 
 import 'package:flutter/scheduler.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  late PerformanceModeHandler handler;
+  late SchedulerBinding binding;
 
   setUpAll(() {
-    handler = PerformanceModeHandler.instance;
+    WidgetsFlutterBinding.ensureInitialized();
+    binding = SchedulerBinding.instance;
   });
 
   test('PerformanceModeHandler make one request', () async {
     const int handle1 = 1;
-    final bool success = handler.createRequest(handle1, DartPerformanceMode.latency);
+    final bool success = binding.createPerformanceModeRequest(handle1, DartPerformanceMode.latency);
     expect(success, isTrue);
-    expect(handler.getRequestedPerformanceMode(), equals(DartPerformanceMode.latency));
+    expect(binding.getRequestedPerformanceMode(), equals(DartPerformanceMode.latency));
 
-    handler.disposeRequest(handle1);
-    expect(handler.getRequestedPerformanceMode(), isNull);
+    binding.disposePerformanceModeRequest(handle1);
+    expect(binding.getRequestedPerformanceMode(), isNull);
   });
 
   test('PerformanceModeHandler make conflicting requests', () async {
     const int handle1 = 1;
-    final bool success1 = handler.createRequest(handle1, DartPerformanceMode.latency);
+    final bool success1 = binding.createPerformanceModeRequest(handle1, DartPerformanceMode.latency);
     expect(success1, isTrue);
 
     const int handle2 = 2;
-    final bool success2 = handler.createRequest(handle2, DartPerformanceMode.throughput);
+    final bool success2 = binding.createPerformanceModeRequest(handle2, DartPerformanceMode.throughput);
     expect(success2, isFalse);
 
-    expect(handler.getRequestedPerformanceMode(), equals(DartPerformanceMode.latency));
+    expect(binding.getRequestedPerformanceMode(), equals(DartPerformanceMode.latency));
 
-    handler.disposeRequest(handle1);
-    expect(handler.getRequestedPerformanceMode(), isNull);
+    binding.disposePerformanceModeRequest(handle1);
+    expect(binding.getRequestedPerformanceMode(), isNull);
   });
-
-
 
   test('PerformanceModeHandler revert only after last requestor disposed', () async {
     const int handle1 = 1;
-    final bool success1 = handler.createRequest(handle1, DartPerformanceMode.latency);
+    final bool success1 = binding.createPerformanceModeRequest(handle1, DartPerformanceMode.latency);
     expect(success1, isTrue);
 
-    expect(handler.getRequestedPerformanceMode(), equals(DartPerformanceMode.latency));
+    expect(binding.getRequestedPerformanceMode(), equals(DartPerformanceMode.latency));
 
     const int handle2 = 2;
-    final bool success2 = handler.createRequest(handle2, DartPerformanceMode.latency);
+    final bool success2 = binding.createPerformanceModeRequest(handle2, DartPerformanceMode.latency);
     expect(success2, isTrue);
 
-    expect(handler.getRequestedPerformanceMode(), equals(DartPerformanceMode.latency));
-    handler.disposeRequest(handle1);
-    expect(handler.getRequestedPerformanceMode(), equals(DartPerformanceMode.latency));
-    handler.disposeRequest(handle2);
-    expect(handler.getRequestedPerformanceMode(), isNull);
+    expect(binding.getRequestedPerformanceMode(), equals(DartPerformanceMode.latency));
+    binding.disposePerformanceModeRequest(handle1);
+    expect(binding.getRequestedPerformanceMode(), equals(DartPerformanceMode.latency));
+    binding.disposePerformanceModeRequest(handle2);
+    expect(binding.getRequestedPerformanceMode(), isNull);
   });
 
 }
